Django-board/board/requirements.txt
absl-py==0.12.0
asgiref==3.4.1
astroid==2.5.1
astunparse==1.6.3
autopep8==1.5.5
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Django==3.2
django-cors-headers==3.11.0
django-etc==1.3.1
django-hitcount==1.3.3
django-session-timeout==0.1.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isort==5.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
mysqlclient==2.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.2.4
pandas-datareader==0.9.0
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.3
wrapt==1.12.1
